backend/requirements.txt
fastapi
uvicorn
sqlalchemy>=2.0
pydantic
psycopg[binary]>=3.1
python-dotenv
